kafka-starter/src/main/java/ru/dev/bachinin/kafka/config/KafkaAutoConfig.java
package ru.dev.bachinin.kafka.config;

import org.apache.kafka.clients.consumer.Consumer;
import org.apache.kafka.clients.consumer.ConsumerConfig;
import org.apache.kafka.clients.consumer.KafkaConsumer;
import org.apache.kafka.clients.producer.KafkaProducer;
import org.apache.kafka.clients.producer.Producer;
import org.apache.kafka.clients.producer.ProducerConfig;
import org.apache.kafka.common.serialization.StringDeserializer;
import org.apache.kafka.common.serialization.StringSerializer;
import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import ru.dev.bachinin.kafka.properties.KafkaProperties;

import java.util.Properties;

@Configuration
public class KafkaAutoConfig {

    @Bean("producerProperties")
    @ConfigurationProperties(prefix = "kafka.producer")
    public KafkaProperties producerProperties() {
        return new KafkaProperties();
    }

    @Bean("consumerProperties")
    @ConfigurationProperties(prefix = "kafka.consumer")
    public KafkaProperties consumerProperties() {
        return new KafkaProperties();
    }

    @Bean("kafkaConsumer")
    @ConditionalOnMissingBean
    public Consumer<String, String> kafkaConsumer(KafkaProperties consumerProperties) {
        Properties props = new Properties();
        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, consumerProperties.getBootstrapServers());
        props.put(ConsumerConfig.GROUP_ID_CONFIG, consumerProperties.getGroupId());
        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
        return new KafkaConsumer<>(props);
    }

    @Bean("kafkaProducer")
    @ConditionalOnMissingBean
    public Producer<String, String> kafkaProducer(KafkaProperties producerProperties) {
        Properties props = new Properties();
        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, producerProperties.getBootstrapServers());
        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
        props.put(ProducerConfig.ACKS_CONFIG, "all");
        return new KafkaProducer<>(props);
    }
}
